refactor: Move AudioPreference & NotificationPreference
